Rename frontend files or components besides `js/src/components` for better consistency in naming conventions.
