+++ js/src/attribute-mapping/attribute-mapping-sync.js
@@ -11,7 +11,7 @@ import { format as formatDate } from '@wordpress/date';
  */
 import usePolling from '.~/hooks/usePolling';
 import { API_NAMESPACE } from '.~/data/constants';
-import glaDateTimeFormat from '.~/utils/date';
+import { glaDateTimeFormat } from '.~/utils/date';
 import HelpPopover from '.~/components/help-popover';
 
 const AttributeMappingSync = () => {
