+++ js/src/product-feed/product-statistics/status-box/account-status.test.js
@@ -7,7 +7,7 @@ import { render } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import REVIEW_STATUSES from '.~/product-feed/review-request/review-request-statuses';
+import { REVIEW_STATUSES } from '../../constants';
 import AccountStatus from '.~/product-feed/product-statistics/status-box/account-status';
 
 jest.mock( '.~/hooks/useAppSelectDispatch' );
