+++ js/src/product-feed/product-statistics/status-box/sync-status.js
@@ -12,7 +12,7 @@ import Status from '.~/product-feed/product-statistics/status-box/status';
 import SyncIcon from '.~/components/sync-icon';
 import SuccessIcon from '.~/components/success-icon';
 import getNumberOfSyncProducts from '.~/utils/getNumberOfSyncProducts';
-import glaDateTimeFormat from '.~/utils/date';
+import { glaDateTimeFormat } from '.~/utils/date';
 
 /**
  * @typedef {import('.~/data/actions').ProductStatistics } ProductStatistics
