+++ js/src/reports/products/products-report-filters.js
@@ -18,7 +18,7 @@ import { getSetting } from '@woocommerce/settings'; // eslint-disable-line impor
 /**
  * Internal dependencies
  */
-import { productsFilter, advancedFilters } from './filter-config';
+import { productsFilter, advancedFilters } from './filterConfig';
 import { recordGlaEvent } from '.~/utils/tracks';
 
 const currency = getSetting( 'currency' );
