+++ js/src/utils/isMobileApp.test.js
@@ -3,7 +3,7 @@
  */
 import { isWCIos, isWCAndroid } from '.~/utils/isMobileApp';
 
-describe( 'isMobile', () => {
+describe( 'isMobileApp', () => {
 	let navigatorGetter;
 
 	beforeEach( () => {
